Put <ResultsTable>-related classes in their own directory.
